refactor: oauth token 엔티티 구조 정리
