+++ src/entities/index.ts
@@ -1,4 +1,5 @@
 export * from "./auth";
 export * from "./connection";
 export * from "./node";
+export * from "./oauth-token";
 export * from "./workflow";
