+++ src/entities/oauth-token/api/connect-oauth-token.api.ts
@@ -1,4 +1,4 @@
-import { request } from "../core";
+import { request } from "@/shared/api/core";
 
 import type { OAuthConnectResponse } from "./types";
