+++ src/entities/oauth-token/api/disconnect-oauth-token.api.ts
@@ -1,4 +1,4 @@
-import { request } from "../core";
+import { request } from "@/shared/api/core";
 
 export const disconnectOAuthTokenAPI = (service: string): Promise<void> =>
   request<void>({
