+++ src/entities/oauth-token/index.ts
@@ -1,0 +1,2 @@
+export * from "./api";
+export * from "./model";
