+++ src/entities/oauth-token/model/index.ts
@@ -1,3 +1,4 @@
+export * from "./query-keys";
 export * from "./useConnectOAuthTokenMutation";
 export * from "./useDisconnectOAuthTokenMutation";
 export * from "./useOAuthTokensQuery";
