+++ src/entities/oauth-token/model/query-keys.ts
@@ -1,0 +1,4 @@
+export const oauthKeys = {
+  all: () => ["oauth"] as const,
+  tokens: () => [...oauthKeys.all(), "tokens"] as const,
+} as const;
