+++ src/shared/api/index.ts
@@ -1,6 +1,5 @@
 export * from "./client";
 export * from "./execution";
-export * from "./oauth-token";
 export * from "./template";
 export * from "./workflow";
 export * from "./query-policy";
