+++ src/shared/model/index.ts
@@ -1,5 +1,4 @@
 export * from "./execution";
-export * from "./oauth";
 export * from "./template";
 export * from "./workflow";
 export * from "./workflowStore";
